requirements.txt
flask
flask-cors
requests
pymupdf
faiss-cpu
numpy
